Move terminal command execution out of registry
