+++ apps/mac/supatermTests/SupatermDebugSnapshotResolverTests.swift
@@ -1,5 +1,4 @@
 import Foundation
-import SupatermSocketFeature
 import Testing
 
 @testable import SupatermCLIShared
